Add serialize-by-type
This is cumbersome but necessary. Types defined in `clj-odbp.types` are not
easily usable in `clj-odbp.serialize.record`.

The factory methods help creating new instances where needed, but we need access
to serialize of both custom and regular types.

This is the only solution I've been able to find so far. It may be better to
check on this and see if a more elegant way is viable.