Add unit tests for base and custom Orient types
It is impossible to have two protocols with a method of the same name in the
same namespace, so I had to add a different method.

See Stuart Sierra answer here: https://groups.google.com/forum/#!topic/clojure/KqWNLkWnw38

Another quirkiness: due to how extend-type works, if I define the custom orient
types in a namespace and use extend-type in a different one, whatever is added
by extend-type is not "seen" when requiring the namespace where the types where
declared.